% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eikos.R
\name{eikos}
\alias{eikos}
\title{Generic method for creating an eikosogram}
\usage{
eikos(y, ...)
}
\arguments{
\item{y}{Either the name of a variable in the data set (eikos.default), or a formula of such variables (eikos.formula).}

\item{...}{includes a data argument as well as other arguments for features of the display}
}
\description{
Generic method for creating an eikosogram
}
\seealso{
\code{\link{eikos.default}}

\code{\link{eikos.formula}}
}
