% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eikos_data.R
\name{eikos_data}
\alias{eikos_data}
\title{Create eikosogram data frame}
\usage{
eikos_data(y, x, data, marginalize = NULL)
}
\arguments{
\item{y}{response variable.}

\item{x}{conditional variables.}

\item{data}{data frame or table to be converted.}

\item{marginalize}{varibale to marginalize on, NULL if none.}
}
\description{
Eikos helper function used to convert data.
}
