% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eikos_legend.R
\name{eikos_legend}
\alias{eikos_legend}
\title{Create eikosogram legend}
\usage{
eikos_legend(labels, colours, margin = unit(2, "points"), lcol = "black")
}
\arguments{
\item{labels}{labels to be included in legend}

\item{colours}{colours}

\item{margin}{unit specifying margin between legend entries}

\item{lcol}{line colour}
}
\description{
Eikos helper function used to create legend.
}
