% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eikos_labels.R
\name{eikos_x_labels}
\alias{eikos_x_labels}
\title{eikos helper function. Returns grob with x axis labels.}
\usage{
eikos_x_labels(x, data, margin = unit(10, "points"), axis_name_size = 12,
  lab_rot = 0)
}
\arguments{
\item{x}{vector of conditional variables}

\item{data}{data frame from eikos_data.}

\item{margin}{unit specifying margins}

\item{axis_name_size}{font size for axis names (points)}

\item{lab_rot}{integer indicating the rotation of the label, default is horizontal}
}
\value{
gList with x labels and x-axis names as grob frames.
}
\description{
eikos helper function. Returns grob with x axis labels.
}
