% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eikos_labels.R
\name{eikos_y_labels}
\alias{eikos_y_labels}
\title{eikos helper function. Returns grob with y axis labels.}
\usage{
eikos_y_labels(y, data, margin = unit(2, "points"), yname_size = 12,
  yvals_size = 10, lab_rot = 0)
}
\arguments{
\item{y}{response variable}

\item{data}{data frame from eikos_data.}

\item{margin}{unit specifying margin}

\item{yname_size}{font size for y axis variable names (in points)}

\item{yvals_size}{font size of labels for values of y variable (in points)}

\item{lab_rot}{integer indicating the rotation of the label, default is horizontal}
}
\value{
gList with x labels and x-axis names as grob frames.

grobFrame with response variable labels and axis text
}
\description{
eikos helper function. Returns grob with y axis labels.
}
