% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eikos_probs.R
\name{eikos_y_probs}
\alias{eikos_y_probs}
\title{Create grob with eikosogram y-axis probabilities}
\usage{
eikos_y_probs(data, yprobs, yprobs_size = 8, margin = unit(2, "points"))
}
\arguments{
\item{data}{data frame from eikos_data object}

\item{yprobs}{vector of probabilities to be shown. NULL if they should be calculated from the data.}

\item{yprobs_size}{font size of labels for horizontal probabilities (in points)}

\item{margin}{unit specifying margin between y axis and eikosogram}
}
\value{
textGrob with y-axis probabilities.
}
\description{
Creates y axis grob to be placed on eikosogram. Called by eikos functions.
}
