% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eikos_legend.R
\name{eikos_legend}
\alias{eikos_legend}
\title{Create eikosogram legend}
\usage{
eikos_legend(labels, title = NULL, yname_size = 12, yvals_size = 12, col,
  margin = unit(2, "points"), lcol = "black")
}
\arguments{
\item{labels}{labels to be included in legend}

\item{title}{if non-NULL a string to give as the legend title}

\item{yname_size}{font size of vertical axis names (in points)}

\item{yvals_size}{font size of labels for values of y variable (in points)}

\item{col}{colours od}

\item{margin}{unit specifying margin between legend entries}

\item{lcol}{line colour}
}
\description{
Eikos helper function used to create legend.
}
